Fix trailing semicolon in dkim for ovh provider
Pulling dns records from ovh to a yaml file puts a semicolon at the end.
Pushing from yaml to ovh will fail the "dkim-compliant" verification if
there is an empty field (and there is one in case of a trailing
semicolon). With the current logic, pulling dkim record created with ovh
then pushing it back will NOT work.

This small patch ignores all empty fields in a dkim records during dkim
validation.